More rpc macros to aseparate modules
