+++ module-system/sov-modules-macros/src/dispatch/mod.rs
@@ -1,7 +1,6 @@
 pub mod cli_parser;
 pub(crate) mod common;
 pub mod default_runtime;
-pub mod derive_rpc;
 pub mod dispatch_call;
 pub mod genesis;
 pub mod message_codec;
