+++ module-system/sov-modules-macros/src/lib.rs
@@ -1,6 +1,7 @@
 #![feature(log_syntax)]
 mod dispatch;
 mod module_info;
+mod rpc;
 use crate::dispatch::cli_parser::CliParserMacro;
 use dispatch::{
     default_runtime::DefaultRuntimeMacro, dispatch_call::DispatchCallMacro, genesis::GenesisMacro,
@@ -141,7 +142,7 @@ pub fn codec(input: TokenStream) -> TokenStream {
 #[proc_macro_attribute]
 pub fn rpc_gen(attr: TokenStream, item: TokenStream) -> TokenStream {
     let input = parse_macro_input!(item as syn::ItemImpl);
-    handle_macro_error(dispatch::derive_rpc::derive_rpc(attr.into(), input).map(|ok| ok.into()))
+    handle_macro_error(rpc::rpc_gen(attr.into(), input).map(|ok| ok.into()))
 }
 
 fn handle_macro_error(result: Result<proc_macro::TokenStream, syn::Error>) -> TokenStream {
@@ -150,21 +151,13 @@ fn handle_macro_error(result: Result<proc_macro::TokenStream, syn::Error>) -> To
         Err(err) => err.to_compile_error().into(),
     }
 }
-/// TODO: this isn't needed anymore
-/// This proc macro generates the actual implementations for the trait created above for the module
-/// It iterates over each struct
-// #[proc_macro_derive(rpc)]
-// pub fn rpc_impls(input: TokenStream) -> TokenStream {
-//     let input = parse_macro_input!(input);
-//     handle_macro_error(dispatch::derive_rpc::rpc_impls(input))
-// }
 
 /// This proc macro generates the actual implementations for the trait created above for the module
 /// It iterates over each struct
 #[proc_macro_attribute]
 pub fn expose_rpc(attr: TokenStream, item: TokenStream) -> TokenStream {
     let input = parse_macro_input!(item as syn::ItemImpl);
-    handle_macro_error(dispatch::derive_rpc::rpc_outer_impls(attr.into(), input))
+    handle_macro_error(rpc::expose_rpc(attr.into(), input))
 }
 
 #[proc_macro_attribute]
