+++ module-system/sov-modules-macros/src/rpc/mod.rs
@@ -1,0 +1,5 @@
+mod expose_rpc;
+mod rpc_gen;
+
+pub(crate) use expose_rpc::expose_rpc;
+pub(crate) use rpc_gen::rpc_gen;
